TAB-462 make "tab helper" configurable (extra content script and app feature)
