+++ src/models/AppFeature.ts
@@ -41,7 +41,8 @@ export enum FeatureIdent {
   ADVANCED_TAB_MANAGEMENT = "ADVANCED_TAB_MANAGEMENT",
   ANALYSE_TABS = "ANALYSE_TABS",
   TAB_GROUPS = "TAB_GROUPS",
-  MONITORING = "MONITORING"
+  MONITORING = "MONITORING",
+  TAB_HELPER = "TAB_HELPER"
 }
 
 export enum FeatureType {
